Ignore "done" on talling Task::take
The check was introduced with timeouts support. But as it turned out it breaks server tasks,
like file transfer for example and many others. So after disconnect they stop working.
Since for server tasks this attribute is nothing. And other regular tasks anyway check
iq.id, the "done" check was dropped as useless. It's a very little performance hit though.